[CMake] Fix LLVM_USE_LLD=On builds (#16611)
Currently we use an older (circa 2021) ld.lld for builds. Since then,
LLD started treating undefined symbols in version scripts as errors.
This means that we cannot use an xmain ld.lld to build xmain.

This change links omptarget with "--undefined-version" if the linker is
LLD, keeping the same behavior as before. (I.e., no error.)