Resolve of merge (conflict) 32ec6d91a120 [SCEV] Make invalidation in SCEVCallbackVH more thorough (.. 'main' -> 'xmain-web' (#16797)
CONFLICT (content): Merge conflict in
llvm/test/Transforms/IndVarSimplify/elim-extend.ll

  commit 32ec6d91a120334d94c1a36764f0bf82a57bb3bb
  Author: Nikita Popov <npopov@redhat.com>
  Date:   Tue Oct 10 10:55:57 2023 +0200

      [SCEV] Make invalidation in SCEVCallbackVH more thorough (#68316)

When a SCEVCallbackVH is RAUWed, we currently do a def-use walk and
remove dependent instructions from the ValueExprMap. However, unlike
      SCEVs usual invalidation, this does not forget memoized values.

The end result is that we might end up removing a SCEVUnknown from the
map, while that expression still has users. Due to that, we may later
fail to invalide those expressions. In particular, invalidation of loop
      dispositions only does something if there is an expression for the
      value, which would not be the case here.

Fix this by using the standard forgetValue() API, instead of rolling a
      custom variant.

      Fixes https://github.com/llvm/llvm-project/issues/68285.

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Fine, Gregory <gregory.fine@intel.com>