[CMPLRLLVM-52337][HIRNonPerfectNestLoopBlocking] Remove an assertion (#16771)
Remove one assertion, which is a conservative change. When the
representative ref returned from "getRepRef" is null, the caller of the
function isValidInnermostLoop() returns false and the transformation
doesn't happen.

Strictly speaking, this PR is not an NFC, but I could not find a test
case.
The issue with assertion was found by loop distribution change, which
introduced a chunk with no memrefs and no liveout temp.