[VPlan][CostModel] TLI corrections to handle cpu-features correctly. (#16780)
Corrected several TLI interfaces and their calls to handle correctlly
"required cpu feature" on vector variant. In those interfaces/calls
the TTI was missed. Also, it appears we have an issue with
CallVecDecisions/SVA during cost modeling. We use the same VPlan to
model different VFs, including peel and remainder loops. When the main
loop is evaluated for one VF but non-masked remainder for another, the
CallVecdecision/SVA are not properly reset. The scenario is: we evaluate
main loop for VF=8 for variant with profitable peeling. The same vplan
is used to evaluate non-masked remainder with VF=4, the
SVA/CallVecDecision are recalculated for VF=4. But we use the already
created CM for the main loop, to calculate cost w/o peeling. And the
call to getCost() does not recalculate CallVecDecision/SVA, so using
results for VF=4 instead of those for VF=8.

https://jira.devtools.intel.com/browse/CMPLRLLVM-52366