CMPLRLLVM-52455: Extend removeRedundantNodes utility to update number of exits for multi-exit loops (#16816)
Some transformations (e.g. complete unrolling) calls
removeRedundantNodes
to clean the HIR and remove nodes that are redundant (e.g. if (1 > 2)).
The problem is if there is a loop set as multi-exit, and the GOTO nodes
are removed, then the loop stops being multi-exit but the information
is not updated. This patch adds updateNumLoopExits after visiting a loop
in removeRedundantNodes to update the loop's multi-exit information.

Also, it adds the call to updateNumLoopExits during unrolling for parent
loops if they are multi-exit loops. This secures that the number of exits are
updated in the parent loops when the inner loop is unrolled.