CMPLRLLVM-51776: Disable HIRLMM if memref doesn't dominate a goto (#16887)
HIRLMM will consider if the write back from the temp to the memref
before a goto requires to be wrapped inside an if (i1 != 0). The problem
is if memref is inside a different parent than the goto, then the check
will be missed and the condition will be added. This creates incorrect
code.

This patch disables HIRLMM if the analysis process finds that a memref
inside the loop happens before a goto, but it doesn't dominates it. In
that case we can't guarantee if the memref will always execute before
the goto.