Resolve of merge (conflict) 088d272e8325 [ADT][DebugInfo][RemoveDIs] Add extra bits to ilist_iterat.. 'main' -> 'xmain-web' (#16955)
CONFLICT (content): Merge conflict in llvm/lib/IR/Instruction.cpp

  commit 088d272e83259a5d8e577a3d2e62012c42a9f9db
  Author: Jeremy Morse <jeremy.morse@sony.com>
  Date:   Tue Oct 17 15:08:58 2023 +0100

[ADT][DebugInfo][RemoveDIs] Add extra bits to ilist_iterator for
debug-info

      ...behind an experimental CMAKE option that's off by default.

This patch adds a new ilist-iterator-like class that can carry two extra
bits
as well as the usual node pointer. This is part of the project to remove
debug-intrinsics from LLVM: see the rationale here [0], they're needed
to
signal whether a "position" in a BasicBlock includes any debug-info
before or
      after the iterator.

This entirely duplicates ilist_iterator, attempting re-use showed it to
be a
false economy. It's enable-able through the existing ilist_node options
interface, hence a few sites where the instruction-list type needs to be
updated. The actual main feature, the extra bits in the class, aren't
part of
the class unless the cmake flag is given: this is because there's a
compile-time cost associated with it, and I'd like to get everything
in-tree
      but off-by-default so that we can do proper comparisons.

Nothing actually makes use of this yet, but will do soon, see the Phab
patch
      stack.

[0]
https://discourse.llvm.org/t/rfc-instruction-api-changes-needed-to-eliminate-debug-intrinsics-from-ir/68939

      Differential Revision: https://reviews.llvm.org/D153777

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Fine, Gregory <gregory.fine@intel.com>