Resolve of merge (conflict) eee8dd90887c [CodeExtractor] Allow to use 0 addr space for aggregate ar.. 'main' -> 'xmain-web' (#16981)
CONFLICT (content): Merge conflict in
llvm/include/llvm/Transforms/Utils/CodeExtractor.h
CONFLICT (content): Merge conflict in
llvm/lib/Transforms/Utils/CodeExtractor.cpp

  commit eee8dd90887cbf86fa0fea1ff770377a87af0257
  Author: Dominik Adamski <dominik.adamski@amd.com>
  Date:   Wed Oct 18 20:12:31 2023 +0200

[CodeExtractor] Allow to use 0 addr space for aggregate arg (#66998)

      The user of CodeExtractor should be able to specify that
the aggregate argument should be passed as a pointer in zero address
      space.

CodeExtractor is used to generate outlined functions required by OpenMP
runtime. The arguments of the outlined functions for OpenMP GPU code
are in 0 address space. 0 address space does not need to be the default
      address space for GPU device. That's why there is a need to allow
the user of CodeExtractor to specify, that the allocated aggregate
parameter
      is passed as pointer in zero address space.

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Fine, Gregory <gregory.fine@intel.com>