[HIRRegionIdentification] Exclude loop preheaders from HIR region whose size exceeds threshold (#16956)
We are not able to clean up outgoing IR generated by HIRCodeGen as well
as the
incoming IR using our current cleanup passes when the incoming IR
contains lots of
opportunities for CSE. Moreover, there are compile time issues
associated with
forming HIR for big basic blocks. For now it is better to throttle such
basic
blocks. We already such a threshold for this in region identification
but it is
not being invoked for non-loop bblocks. This PR adds the size check for
loop
preheaders.

CMPLRLLVM-52664