[HIRParser] Bugfix for incorrectly considering a 'merge numerator' operation as a mathematical add operation (#16982)
We were using CanonExprUtils::add() operation which is mathematical add
operation to execute 'merge numerator' operation. They are different
when one of
the CEs has a non-unit denominator. Since we don't really have a 'merge
numerator'
utility, we simulate it by temporarily setting the denominator to 1
before
calling add() and restoring the denominator afterwards.

CMPLRLLVM-52388