[HIRLoopDistributionMemRec] Perform trip count multiversioning if we decide to do stripmining (#17068)
This is a follow-up on the previous change to trigger distribution for
vectorization. This PR introduces trip count multiversioning to skip
loops with small trip count since stripmining and scalar-expansion may
add unnecessary overhead. It also introduces a check to skip
distribution
of loops which likely have a small trip count and some small code
cleanup.

CMPLRLLVM-52745