[VPlan][CostModel]Correction in spill/fill heuristic (#17046)
The SF heuristic calculates overhead for each basic block
in VPlan but does not account that there are loop blocks and
non-loop blocks (preheader etc). As a result, possible
overhead of non-loops blocks is added to the iteration cost.
This PR adds one more parameter to apply() method of the
heuristics and makes SF overheads accounted separately. Other
heuristics are unchanged.