CMPLRLLVM-25534: Preprocess non-unit stride access pass for HIR (part 1) (#16951)
This change set implements the skeleton for the Preprocess non-unit
stride
access in the HIR. The basic idea is if the vectorizer can't vectorize
the access
to a memory reference, then we copy the data to a local array, vectorize
the loops with the copied data, and then copy back. The accesses in the
local
array is now unit strided, therefore the vectorizer should be able to
kick in.