Resolve of merge (conflict) cee08ff342f3 [Libomptarget] Do not pass 'nogpulib' to the non-LTO Nvidi.. 'main' -> 'xmain-web' (#17157)
CONFLICT (content): Merge conflict in openmp/libomptarget/test/lit.cfg

  commit cee08ff342f39f8cd2b3b66f48ecb33d8b5efe65
  Author: Joseph Huber <35342157+jhuber6@users.noreply.github.com>
  Date:   Thu Oct 26 10:36:34 2023 -0500

[Libomptarget] Do not pass 'nogpulib' to the non-LTO Nvidia tests
(#70327)

      Summary:
For the other tests we pass `-nogpulib` to ensure that we set up the
      needed libraries correctly. However, this caused problems for the
non-LTO build and test of Nvidia systems. In general this is because we
would do a separate compile of the libomptarget device runtime and then
link in that cubin. This exercised the runtime in a lot of ways it's not
used to, since doing things this way was hardly expected or tested. This
patch disables it only for the Nvidia non-LTO build so that we still get
the effect of `--liboimptarget-nvptx-bc-path` rather than ignoring it.

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Fine, Gregory <gregory.fine@intel.com>