[HIRParser] Pick up smaller extent whenever there is a mismatch in extent info in the incoming IR (#17139)
This is related to CMPLRLLVM-52669. In fortran src code, the caller with
a
bigger sized array can pass a slice of the array to the callee. The
callee can declare the same array as a smaller sized array inside the
function. If the callee is inlined into the caller, we see the
mismatched info
in the IR and assert. This prevents FFE from adding extent info in more
cases.

Now, instead of asserting we pick up the more refined, smaller extent.