CMPLRLLVM-52892: Prevent computing number of exists for empty loops while removing redundant nodes (#17156)
The process of removing redundant nodes computes the number of exits
when a loop is multi exit. If a loop will be removed by this process,
then it will be an empty loop since it's children nodes were removed
too. Computing the number of exists asserts if the loop is empty.

This patch adds an extra check in the remove redundant nodes process
to check if the loop will be removed, then don't compute the number of
exits.