[SYCL] Change linkage to linkonce_odr unless SYCL_EXTERNAL. (#10317) (#17189)
Relands upstream change 4cb4fec940cf5 after lost in pulldown
and updates xmain-only tests.

In getLLVMLinkageForDeclarator, CUDA device functions are given internal
linkage to allow improved IPO. The same optimizations would also be
useful for SYCL. However, as multiple translation units are a
possibility here, it cannot be done exactly the same way. There are a
number of extra considerations that need to be made.

- If a device function or object is annotated with the SYCL_EXTERNAL
macro, it may be called from a different module and its handling should
not be touched.

- If a device function or object is not annotated with the SYCL_EXTERNAL
macro, it may not be called from a different translation unit unless
that other translation unit provides its own definition of it. As such,
if other optimizations remove all references in the current module, the
function or object itself can be removed, and other passes should be
aware that it is legal to remove the function or object.

- If two kernels defined in different translation units call into the
same non-kernel inline or template device function, these different
copies of the device function would previously be merged together, and
should continue to be.

The linkage that satisfies these criteria is linkonce_odr, so this
commit marks functions and objects so, unless the sycl_kernel or
sycl_device attributes are present which indicate that it must remain
externally accessible.

Co-authored-by: Harald van Dijk <harald@gigawatt.nl>