[VPlan] Tune peeling decisions using minimum peel TC and estimated TC. (#16813)
This patch adds a new flag `vplan-optimize-profitable-peel-check`
(default: `false`) which, when enabled, causes the following behavior in
the Planner:

1. When the computed minimum profitable peel TC is *greater* than the
estimated max TC, disable peeling, as it is likely that the TC check
will always be false and the peel loop will always be skipped.

2. When the computed minimum profitable peel TC is *lesser* than the
estimated min TC, disable the TC check, as it is likely that the TC
check will always be true and the peel loop will always execute.

Additionally, we now compute a minimum profitable peel TC for all
VPlans. The previous behavior was to only compute this TC when the plan
being considered was better than the previous best; doing so for all
plans enables the aforementioned behavior.

This was seen in local builds to improve performance in 
`polyhedron2011F/air@opt_fast_avx512_ipo_nostdrealloc` by 5%.

JR: CMPLRLLVM-52457.