Update DTrans tests for default data layout that pads structures (#17233)
The original DTrans delete field used a default data layout for
structures that did not include padding because i64 types were
not forced to 8-byte alignment. However, the combination of
adding a target triple to the test to support -intel-libirc-allowed
and the community change to now use the target triple to infer
a data layout has resulted in the need to modify the size
constants used within these tests when analyzing the structure
allocations. In some cases, the structures also needed to be
modified to get a structure of the appropriate size for the
test.

[CMPLRLLVM-53024]