[VPlan][Remainder Evaluator]Correct trip count estimation for remainder (#16897)
There are two issues. 
- The remainder evaluator for unmasked remainder always uses the main
loop VF and UF to calculate cost of remainder even when TC is known.
That is incorrect we should use calculations that we make in constructor
depending on the original TC.
- The remainder TC is calculated incorrectly when the original TC is not
estimated. we should use just main VF and UF instead of default trip
count.

This is a partial fix for
https://jira.devtools.intel.com/browse/CMPLRLLVM-38955