[CMPLRLLVM-53184] Adjust anti-pattern checking (#17239)
CMPLRLLVM-53184


[CMPLRLLVM-48772](https://jira.devtools.intel.com/browse/CMPLRLLVM-48772)
was enabling book example. Still, the book-example was caught as an
anti-pattern (a pattern against loop blocking, unlikely to be profitable
with loop blocking).

This JR is about to adjust the check for such anti-pattern so that book
example is not caught as anti-pattern but to be blocked.

 

NOTE: anti-pattern was introduced for a test in coremark-pro.