[NFC][VPlan] Clarify seemingly-swapped arguments in preprocessPrivateFinalCondInstructions() (#17300)
Coverity complains that the call

    VPBB->setTerminator(VPBBFalse, VPBBTrue, AllZeroCheck);

has its arguments in the wrong order, since the order of
`setTerminator()` should be the `true` target first, then the `false`
target.

It turns out that the code is correct, and Coverity didn't catch a bug,
but the code as-written is a bit confusing. `VPBBFalse` denotes the
branch where the private *is not* stored to, but this is actually the
`true` target of the branch, as the branch condition is whether or not
the mask value is zero. Conversely `VPBBTrue` denotes the branch where
the private *is* stored to, but this is actually the `false` target of
the branch.

This patch rewords the branch target names (`VPBBFalse` ->
`VPBBMaskZero`, `VPBBTrue` -> `VPBBMaskNonZero`) to be more explicit as
to
what is being checked. As an added bonus, these names match ones used in
the example code from the preceding comment.

JR: CMPLRLLVM-53225.