Resolve of merge (conflict) 7d77bbef4ad9 [DebugInfo][RemoveDIs] Add prototype storage classes for ".. 'main' -> 'xmain-web' (#17320)
CONFLICT (content): Merge conflict in
llvm/lib/Transforms/IPO/DeadArgumentElimination.cpp

  commit 7d77bbef4ad9230f6f427649373fe46a668aa909
  Author: Jeremy Morse <jeremy.morse@sony.com>
  Date:   Thu Nov 2 10:36:51 2023 +0000

[DebugInfo][RemoveDIs] Add prototype storage classes for "new"
debug-info

This patch adds a variety of classes needed to record variable location
debug-info without using the existing intrinsic approach, see the
rationale
      at [0].

The two added files and corresponding unit tests are the majority of the
plumbing required for this, but at this point isn't accessible from the
rest of LLVM as we need to stage it into the repo gently. An overview is
that classes are added for recording variable information attached to
Real
(TM) instructions, in the form of DPValues and DPMarker objects. The
metadata-uses of DPValues is plumbed into the metadata hierachy, and a
field added to class Instruction, which are all stimulated in the unit
tests. The next few patches in this series add utilities to convert
to/from
this new debug-info format and add instruction/block utilities to have
debug-info automatically updated in the background when various
operations
      occur.

This patch was reviewed in Phab in D153990 and D154080, I've squashed
them
together into this commit as there are dependencies between the two
      patches, and there's little profit in landing them separately.

[0]
https://discourse.llvm.org/t/rfc-instruction-api-changes-needed-to-eliminate-debug-intrinsics-from-ir/68939

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Fine, Gregory <gregory.fine@intel.com>