[HIRMVForConstUB] Fix to preserve MAX_TC_EST unless loop is converted into constant trip loop (#17305)
This came up as a result of assertion in CMPLRLLVM-53229. The issue is
that
locality analysis asserts when the refs in a sorted group have a
negative
distance between group. This can happen if the subscript goes out of
range. For
example we can have this group-

A[0][i+8]
A[1][i]

If the number of elements in lowest dimension is 4, the refs will have
negative
distance as i+8 goes out of range.

The out-of-range indices can be produced by some loopopt transformations
if
MAX_TC_EST is lost. In this case, we unrolled & jammed a loop by 8 whose
MAX_TC_EST was 6. Ideally speaking, we should not assert on this. It is
possible
for locality analysis to ignore such cases and become conservative but I
have
found this assertion to be a good way of finding out when a
transformation drops
this information. This is probably the 3rd transformation where we
discovered
this issue. I think we should keep the assertion until we encounter a
case where
the issue cannot be resolved easily.