[VPlan][CostModel] Don't allow large VF for calls when compiling for -mprefer-vector-width < 512 (#17137)
The code generator does not support calls with 512-bit vector
parameters, splitting them into ymms when compiling for
AVX512 with -mprefer-vector-width=256, and thus violating
ABI. 
This is a workaround to avoid such vectorization of the calls by
assigning a big cost for them. The workaround is provided due
to complexity of implementation in CG and should be removed
after they fix the issue. This does not apply to svml calls
due to we have a special processing in MapIntrinToIml pass.

https://jira.devtools.intel.com/browse/CMPLRLLVM-52715