[COFF] Enable -llvm-verify-each functionality for LTO phase (#17366)
The verify pass only runs at specific pipeline locations,
leading to JIRAs being assigned to the pass closest to
the verifier run, even if the problem was caused by a
pass much earlier in the pipeline, such as what happened
with CMPLRLLVM-52925. The compile step pipeline supports
an option to clang named, -llvm-verify-each, which enables
verifier runs between transformation passes. This change
enables the link step to have this functionality when
using /opt:llvm-verify-each functionality.

[CMPLRLLVM-53333]