"CMPLRLLVM-53170 Reroll bugs in self SR with many operands" (#17361)
- [CMPLRLLVM-53170][HIRLoopReroll] Fixing bugs in handling a self SR
with multiple operands.

The immediate assertion failure was from the lack of makeConsistent.
Blob was not correctly attached to a regular ddref after rerolling.
Resulting reroll was also wrong. Reroll's logic for a self reduction
inst was not robust enough with multiple operands. (e.g. S = S + a + b +
c + d). This PR is fixing the those bugs.