[VPlan][EarlyExitVec] Add HIR vector codegen support for explicit early-exit loops. (#16603)
This PR updates VPlan HIR vectorizer codegen to support the new
instructions introduced to represent early-exit loops explicitly in
VPlan. Minor corrections are also made in several parts of codegen and
VPlan pipeline to not confuse explicit early-exit loops with search-loop
idiom that is treated specially. With the changes added in this PR, the
prototype for explicit early-exit loop vectorization feature will be
functional for end-to-end and regression testing.

JIRA: CMPLRLLVM-50417
Change-Id: Iafbe0be1461c559585e40e990bbb96d47b2cbc7c