Make sure null ptr ref is created as a constant '0' CanonExpr by createConstDDRef() (#17401)
The null ptr ref was being created using 'ConstantPointerNull' value as
a blob
by DDRefUtills::createConstDDRef(Value *). This was not expected by
Vectorizer
as the framework usually represents it as a '0' CanonExpr and there is
no utils
interface to get null value blob. For consistency sake, I have updated
the
utility to behave the same way as the framework and also added
verification code
to catch such an issue in the future.

CMPLRLLVM-52919