CMPLRLLVM-52742: Backend analysis for Fortran dope vector element types (#17403)
The Fortran front end implementation is checked into the fort branch and
will be promoted next week. This is the analysis pass needed to make the
results available to transformation passes that need it.

The result is to provide two utility functions:
  isDopeVectorType(Type *)
  getDopeVectorPtrType(Type *)
for the transformation passes that need them.

Once this is checked in, I can start replacing the calls to the old
routines taht use pattern matching of structures and strings in the
types to get this information.