[VPlan][ValueTracking] Account for outer IVs when computing KnownBits. (#17430)
In the HIR path, when we compute KnownBits for a VPlanSCEV, we do so by
analyzing the `CanonExpr` representing the base of the induction.
Previously, this was done by summing the KnownBits for the constant
value and the KnownBits for its blobs, but this is only partially
correct -- this ignores the possibility that the CE contains an IV. This
case actually occurs when computing KBs for address SCEVs for memory
references with an outer IV, e.g:

    + DO i1 = 0, %n, 1
    |   + DO i2 = 0, %n, 1
    |   |   (@A)[4 * i1 + i2] = 0;
    |   + END LOOP
    + END LOOP

In the above, we would compute KnownBits for the induction representing
`(@A)[4*i1 + i2]` as if the base were `(@A)[0]` rather than
`(@A)[4*i1]`. In the event that `@A` is not unknown (e.g is aligned),
this may cause us to assume that memrefs like the above are actually
more aligned than they are. This causes cost inaccuracies when modeling
load/store costs, and also blocks the enablement of general alignment
propagation in HIR (#17045).

This patch fixes the above issue by adding logic to compute KBs for IVs
that may be contained in the base CE for a VPlanSCEV in the HIR path.

JR: CMPLRLLVM-53384.