Resolve of merge (conflict) f1b0a544514f Reapply 7d77bbef4ad92, adding new debug-info classes 'main' -> 'xmain-web' (#17472)
CONFLICT (content): Merge conflict in
llvm/lib/Transforms/IPO/DeadArgumentElimination.cpp

  commit f1b0a544514f3d343f32a41de9d6fb0b6cbb6021
  Author: Jeremy Morse <jeremy.morse@sony.com>
  Date:   Wed Nov 8 14:58:34 2023 +0000

      Reapply 7d77bbef4ad92, adding new debug-info classes

      This reverts commit 957efa4ce4f0391147cec62746e997226ee2b836.

      Original commit message below -- in this follow up, I've shifted
un-necessary inclusions of DebugProgramInstruction.h into being forward
declarations (fixes clang-compile time I hope), and a memory leak in the
      DebugInfoTest.cpp IR unittests.

I also tracked a compile-time regression in D154080, more explanation
there, but the result of which is hiding some of the changes behind the
EXPERIMENTAL_DEBUGINFO_ITERATORS compile-time flag. This is tested by
the
      "new-debug-iterators" buildbot.

[DebugInfo][RemoveDIs] Add prototype storage classes for "new"
debug-info

This patch adds a variety of classes needed to record variable location
debug-info without using the existing intrinsic approach, see the
rationale
      at [0].

The two added files and corresponding unit tests are the majority of the
plumbing required for this, but at this point isn't accessible from the
rest of LLVM as we need to stage it into the repo gently. An overview is
that classes are added for recording variable information attached to
Real
(TM) instructions, in the form of DPValues and DPMarker objects. The
metadata-uses of DPValues is plumbed into the metadata hierachy, and a
field added to class Instruction, which are all stimulated in the unit
tests. The next few patches in this series add utilities to convert
to/from
this new debug-info format and add instruction/block utilities to have
debug-info automatically updated in the background when various
operations
      occur.

This patch was reviewed in Phab in D153990 and D154080, I've squashed
them
together into this commit as there are dependencies between the two
      patches, and there's little profit in landing them separately.

[0]
https://discourse.llvm.org/t/rfc-instruction-api-changes-needed-to-eliminate-debug-intrinsics-from-ir/68939

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Wang, Xin10 <xin10.wang@intel.com>