[Cov] Extend SPIEmitter pass to write data structures to file (#17486)
This is part of the extension to the code coverage usage model
to have a single .spi file. This extends the SPIEmitter pass
to write the code coverage data structures into a file format
that is recognizable by the llvm-cov tool. In future changes,
this will be extended to bundle the temporary file into a tar
or zip file, and the llvm-cov tool will be extended to read
the individual sections of the bundled .spi file.

[CMPLRLLVM-52501]