[SYCLTransforms] Add convergent attr recursively to tid calls and users (#17435)
9898e65 removes memory(none) attr from tid and its user functions/calls.
It prevents DCE of unused tid builtin call and thus changes behavior of
max-WG-dim deduction for tests in CMPLRLLVM-53361. The tests are now
vectorized which is not necessary.

This PR solves the hoisting issue in CMPLRLLVM-52890 and CMPLRLLVM-48583
by adding convergent attr to the functions and restores 1eb2569 which is
needed for vector code.
Run LinearIdResolverPass before AddFunctionAttrsPass so that the latter
needs not to handle linear tid builtins.

An alternative approach could be adding readonly attribute to tid and
its user functions/calls. However, a concern is that LICM is still able
to hoist readonly calls in certain cases we haven't seen.