[CMPLRLLVM-53316] Bailout DD for fusion is source and destination loops' UBs have non-constant distance. (#17417)
In case of following loop sequence:
  for (i =0; i<n ;i++) {
    for (j =0;j< 2*n; j++) {
      a[j][i] = j;
    }
  }
  for (i =0; i<n ;i++) {
    for (j =0;j< n; j++) {
      b[j][i] = a[j+n][i];
    }
  }
DD used to refine the edge between a[j][i] and [j+n][i] as independent
because distance in first dimension is greater than second loop upper
bound, which is incorrect. We can run MIV test in ForFusion mode if UBs
of the Src and Dst loops have constant distance.