CMPLRLLVM-53403: Disable HIR opt-var predicate when inner loops are multi-exit (#17545)
We currently don't process multi-exit loops during HIR opt-var
preducate. The issue is that if there are nested loops, and the inner
loop is multi-exit, then we will try to apply the optimization for a
condition that relates to the outer loop.

This patch disables visiting inner loops that are unknown or multiexit
while analyzing a loop nest.