[EarlyExitVec][ParVecAnalysis] Drop the bailout for early-exit loops with non-unique exit blocks. (#17550)
During development of early-exit loop vectorization feature in VPlan, we
decided to promote entity finalization from exit blocks to the common
merged exit block. This enables us to support loops with non-unique exit
blocks since finalization code will always be executed. This PR drops
the stale bailout check in ParVecAnalysis. This is needed to enable the
early-exit loop vectorization feature in some of the end-to-end tests in
ParoptVecC test suite.

Change-Id: Iadfc8f7dad7eba9d4475a1beb9741b74d5e9471d