Fix ocl32-bit build on Windows (#17667)
There was a place where we were passing a SmallVectorImpl by value in a
function that is never called. The caused a build failure when the
SmallVectorImpl destructor was made protected. The fix is to pass the
SmallVectorImpl by reference (which is the intended use of this class).

This fixes CMPLRLLVM-53694