[VPlan][VLS][HIRCG] Changes to avoid optimized RegDDRef creation when dealing with a GEP with constant index (#17637)
Vector code generation for a GEP was trying to fold the constant GEP index into an already created
RegDDRef to avoid doing this through an assignment to a temp. However, this code was ignoring
that the RegDDRef's element type and GEP's result element type may not match causing bad code
to be generated. The fix avoids this optimization when these types do not match.